Consolidate all global constants to a single file
