+++ src/game/raceState.ts
@@ -1,6 +1,6 @@
+import { RACE_LENGTH, SERVER_TICK_RATE_MS } from "../config/globalsettings.js"
 import { InternalHorse } from "./horse/horse.js"
-import { RACE_DURATION, Race } from "./race.js"
-import { SERVER_TICK_RATE_MS } from "./gameServer.js"
+import { Race } from "./race.js"
 
 export class RaceState {
     horseStates: Array<HorseState> = []
@@ -18,7 +18,7 @@ export class RaceState {
 
     nextState(): RaceState {
         let next = new RaceState()        
-        next.time = this.time + SERVER_TICK_RATE_MS 
+        next.time = this.time + SERVER_TICK_RATE_MS
         next.horseStates = this.horseStates.map((hs) => {
             if (hs.finishTime !== null) { return hs }
             let nextHs = new HorseState(hs.horse)
@@ -26,8 +26,8 @@ export class RaceState {
                 Math.min(hs.currentSpeed + hs.horse.acceleration,
                          hs.horse.topSpeed)
             nextHs.position = hs.position + hs.currentSpeed
-            if (nextHs.position > RACE_DURATION) {
-                nextHs.position = RACE_DURATION
+            if (nextHs.position >= RACE_LENGTH) {
+                nextHs.position = RACE_LENGTH
                 nextHs.finishTime = next.time
             }
